Refactored some code and divided into more specific folders
