+++ src/components/HomePage/HomePage.module.css
@@ -26,11 +26,11 @@
 
 .mainLogo {
     text-align: center;
-    padding-bottom: 8em;
+    padding-bottom: 3em;
 }
 
 .mainLogo img {
-    width: 40%;
+    width: 50%;
 }
 
 .infoBox {
