+++ src/components/HomePage/SearchArea/SearchArea.module.css
