+++ src/components/HomePage/SearchArea/SearchBar.js
