+++ src/components/HomePage/SearchArea/SliderValueTooltip.js
@@ -1,0 +1,37 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+import Tooltip from '@material-ui/core/Tooltip';
+
+const SliderValueTooltip = props => {
+    const { children, open, value } = props;
+  
+    const popperRef = React.useRef(null);
+    React.useEffect(() => {
+        if (popperRef.current) {
+            popperRef.current.update();
+        }
+    });
+  
+    return (
+        <Tooltip
+            PopperProps={{
+                popperRef,
+            }}
+            open={open}
+            enterTouchDelay={0}
+            placement="bottom"
+            title={value}
+        >
+            {children}
+        </Tooltip>
+    );
+};
+  
+SliderValueTooltip.propTypes = {
+    children: PropTypes.element.isRequired,
+    open: PropTypes.bool.isRequired,
+    value: PropTypes.number.isRequired,
+};
+
+export default SliderValueTooltip;
