+++ src/components/HomePage/SearchArea/jobTypes.js
