+++ src/components/HomePage/SearchResultsArea/OfferCard.js
