+++ src/components/HomePage/SearchResultsArea/OfferCard.module.css
