+++ src/pages/HomePage.js
@@ -1,41 +1,38 @@
-import React, { Component } from 'react';
+import React from 'react';
 
 import MainView from "../components/HomePage/MainView";
-import SearchResults from "../components/HomePage/SearchResults";
-import ProductDescription from "../components/HomePage/ProductDescription";
+import SearchResults from "../components/HomePage/SearchResultsArea/SearchResults";
+import ProductDescription from "../components/HomePage/ProductPlacementArea/ProductDescription";
 
+function HomePage() {
+    
+    let [productDescriptionRef, setProductDescriptionRef] = React.useState(null);
+    let [searchResultsRef, setSearchResultsRef] = React.useState(null);
 
-class HomePage extends Component {
-    constructor(props) {
-        super(props);
-
-        this.state = {
-            productDescriptionRef: null
-        };
-    }
-
-    render() {
-        return (
-            <React.Fragment>
-                <MainView scrollToProductDescription={this.scrollToProductDescription}/>
-                <ProductDescription setRef={this.setProductDescriptionRef}/>
-                <SearchResults />
-            </React.Fragment>
-        );
-    }
-
-    setProductDescriptionRef = ref => {
-        this.setState({
-            productDescriptionRef: ref
-        });
-    }
-
-    scrollToProductDescription = () => {
-        this.state.productDescriptionRef.current.scrollIntoView({
+    const scrollToProductDescription = () => {
+        productDescriptionRef.current.scrollIntoView({
             behavior: "smooth",
             block: "start"
         });
-    }
+    };
+    const scrollToSearchResults = () => {
+        searchResultsRef.current.scrollIntoView({
+            behavior: "smooth",
+            block: "start"
+        });
+    };
+
+    return (
+        <React.Fragment>
+            <MainView 
+                scrollToProductDescription={scrollToProductDescription}
+                scrollToSearchResults={scrollToSearchResults}
+            />
+            <ProductDescription setRef={setProductDescriptionRef}/>
+            <SearchResults setRef={setSearchResultsRef}/>
+        </React.Fragment>
+    );
+    
 }
 
 export default HomePage;
